requirements.txt
Django==1.11.12
argparse==1.2.1
certifi==2018.4.16
chardet==3.0.4
django-cors-headers==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.6
pytz==2018.4
requests==2.18.4
urllib3==1.22
wsgiref==0.1.2
